requirements.txt
Django==3.1.2
psycopg2==2.8.6
django-cors-headers
django-allauth
djangorestframework
sqlparse==0.4.1
pillow
django-filter
django-mptt
dj_database_url
gunicorn
whitenoise
python-dotenv
channels
pika
aio_pika
django-organizations
django_extensions
./django-rest-auth
./django-rest-framework-jwt
pylint_django
pytest-django
pytest-subtests
drf-extensions
django-safedelete
factory-boy